# Distributor: Reuse allocation on line truncation