**What this PR does / why we need it**:
Reuses allocation on calls to `truncateLines` and add a guard-clause to check if `maxLineSize==0` out of the loop.
It also adds the number of validated log lines to the existing `finished to validate request` event.

Before:
```
Running tool: /usr/local/go/bin/go test -benchmem -run=^$ -tags integration,requires_docker -bench ^BenchmarkTruncateLogLines$ github.com/grafana/loki/pkg/distributor

goos: linux
goarch: amd64
pkg: github.com/grafana/loki/pkg/distributor
cpu: AMD Ryzen 9 7900X 12-Core Processor            
BenchmarkTruncateLogLines-24    	  139802	      8536 ns/op	      64 B/op	       2 allocs/op
PASS
ok  	github.com/grafana/loki/pkg/distributor	1.302s
```

After:
```
Running tool: /usr/local/go/bin/go test -benchmem -run=^$ -tags integration,requires_docker -bench ^BenchmarkTruncateLogLines$ github.com/grafana/loki/pkg/distributor

goos: linux
goarch: amd64
pkg: github.com/grafana/loki/pkg/distributor
cpu: AMD Ryzen 9 7900X 12-Core Processor            
BenchmarkTruncateLogLines-24    	  151281	      8118 ns/op	      64 B/op	       2 allocs/op
PASS
ok  	github.com/grafana/loki/pkg/distributor	1.330s
```